# backend/swagger.yaml
openapi: 3.0.3
info:
  title: MovieHub API
  version: 1.0.0
  description: Auth contract between Backend and Mobile.
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token in Authorization header as Bearer token.
    refreshCookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: HttpOnly refresh token cookie set by login/register/refresh.
  schemas:
...
    ApiSuccess:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
      required: [success, data, message]
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue new access token
      security:
        - refreshCookieAuth: []
      responses:
        '200':
          description: Token refresh success
          headers:
            Set-Cookie:
              description: Sets rotated HttpOnly refreshToken cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '401':
          description: Invalid or missing refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...